refactor(web): extract update checker into its own module
Move UpdateStatus / UpdateCommit types, updateStatusCache, currentGitHead,
parseGitHubRemote, refreshUpdateStatus and startUpdateChecker into
src/web/update-checker.ts. Cache is now module-private; routes read it
via getUpdateStatus().

web.ts: 4091 -> 3980 lines.
